[Fix]: Button commands breakage with `help` (#10658)
As part of refactor to remove passing explicit width prop down to each element (see PR
#[10342](https://github.com/streamlit/streamlit/pull/10342)), some breakage occurred in button commands (`st.button`, `st.link_button`, `st.download_button`, `st.form_submit_button`, etc.) when using the `help` param.

* `Tooltip.tsx`: we pass a prop, `containerWidth`, down to the tooltip wrapper div so that if `containerWidth` is true, its width is `"100%"`, otherwise set width to `"auto"`. Previously we always set the width to 100%. This prop is passed from button command > `BaseButtonTooltip` > `TooltipIcon` > `Tooltip`.

We used to use a `fluidWidth` concept when `useContainerWidth` was true & using `help`, which required passing the container width down to the button to render properly. Since we are no longer passing width down, simplifying by renaming to `containerWidth` to control rendering width at `100%` or `auto`.