Fix support for dask data objects (#10662)
## Describe your changes

Dask recently removed the legacy types for dataframe, series, and
object. This broke our support for dask objects and caused our
integration tests to fail. This PR adds support for the new types.

## Testing Plan

- Existing tests should cover the new types. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.