Add links for Google and ChatGPT to exceptions (#10078)
## Describe your changes

Adds some small links at the bottom of every exception to open Google or
ChatGPT with the error message (and for ChatGPT also the stack trace)
inserted into the query.



https://github.com/user-attachments/assets/c6b5544d-95e7-42f5-b0e4-3dd749a81ac5

Note that currently, the links are also added to exceptions shown via
`st.exception`. I'm not sure if that's useful or annoying. But given
that `st.exception` is not used very often and we can change this
anytime without breaking, I don't think it's worth implementing anything
right now.

## GitHub Issue Link (if applicable)

## Testing Plan

- Updated existing snapshot tests. I think that should cover it well
enough.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>