Popover with help and use_container_width is rendered correctly (#10709)
## Describe your changes

This fixes an issue where `use_container_width` is ignored if the `help`
parameter is passed to `st.popover` by ensuring
`element.useContainerWidth` is provided to `BaseButtonTooltip` as well
as `BaseButton`.

## GitHub Issue Link (if applicable)

Closes https://github.com/streamlit/streamlit/issues/10693

## Testing Plan

Adds some playwright tests similar to the existing `use_container_width`
tests but including the help option.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.