Allow user to choose whether to print error finding secrets  (#10508)
## Describe your changes

Currently, we both raise an error and print the error in the app when
there is a problem loading the secrets from the toml file.

This change removes the printing so that the app developer can decide
whether to display the error to the end user. The Streamlit customized
message for the error is added to the raised exception for debugability.

Image with no error handling in the app code:

![Screenshot 2025-02-25 at 4 11
11 PM](https://github.com/user-attachments/assets/69881787-ba0b-4cf3-b1b3-4247d5564b80)

 

## GitHub Issue Link (if applicable)

https://github.com/streamlit/streamlit/issues/8559

## Testing Plan

Existing tests cover this change and are updated. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.