Add `st.badge` and Markdown badges (#10090)
## Describe your changes

Adds two ways to add badges to Streamlit:
- A dedicated command `st.badge` that takes label, icon, and color. This
is just a small wrapper around `st.markdown`.
- A new Markdown directive, e.g. `:blue-badge[Badge text]`. 

Note that the `rainbow` color is not supported for badges because I
couldn't get it to work after trying a bunch of different approaches,
and it didn't seem worth investing more time into this given that it
should be a very rare use case. See comment in the code.

![Screenshot 2025-03-11 at 23 49
49@2x](https://github.com/user-attachments/assets/ff3f785d-73af-4a35-8d74-f358ca95c1ff)



## GitHub Issue Link (if applicable)

Closes #6443

## Testing Plan

- Added a quick Python unit test for `st.badge`. 
- Added a small E2E test for `st.badge` and Markdown badges. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>