Add a GitHub comment for the `pr-preview` job (#10700)
## Describe your changes

This PR improves the `pr-preview` job by commenting on the PR with a
link to the wheel feel and deploy preview, similar to Netlify's deploy
preview. Just check out the generated comment below!

![Screenshot 2025-03-09 at 23 38
26](https://github.com/user-attachments/assets/16f616ef-3d95-4123-9ce6-5c13cb502e30)

Specifically, it:
- Creates a basic comment right when the jobs starts to notify the user
that the PR preview is building.
- When the job is done, it updates that comment to say that the PR
preview is ready and includes the link to the wheel file on S3 and the
deploy link for the core preview app.
- Whenever the job runs again (e.g. on a new commit), it updates the
comment again to notify the user of the status.


## GitHub Issue Link (if applicable)

## Testing Plan

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>