Fix usage of "sans serife" as font (#10789)
## Describe your changes

Our old `font` property allowed configuring the font family to `sans
serif`. This isn't a valid CSS font family, but we still want to support
it for backwards compatibility reasons. This PR fixes the mapping that
we apply to map to our font family we defined for `sans-serif`.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/10786

## Testing Plan

- Added unit test to cover this.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.