Fix `createTheme` overriding active theme's baseweb props (#10855)
## Describe your changes

Added logic to create a deep copy of the Base Web UI theme to prevent
modifications to the original theme object. We observed that when
calling createTheme for sidebar, the basewebTheme for main app was being
overridden by the createBaseUiTheme function.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.