Update pytest to latest version (#10846)
## Describe your changes

I tried the same update some time ago in
https://github.com/streamlit/streamlit/pull/9258
But had to revert because of some kind of issue it introduced into our
playwright test: https://github.com/streamlit/streamlit/pull/9942
Unfortunately, we never found out the root cause for the issue.

However, the issue seems to be fixed with the latest version pytest (or
some dependency). I tried lots of reruns of our e2e job without any
indication of this issue.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.