[Fix] Being unable to determine the form a current element is in (#10752)
Fixes #8761

## Describe your changes
When an element is created with dg._block() the _form_data will be
propagated from its parent, but when an element is created with a
dg._enqueue it will not. This makes it so the current code to find a
form will not be able to find the right form, so the easiest way is just
making _enqueue propagate a form. ( current code to find the form: )

https://github.com/streamlit/streamlit/blob/4d0013c5b66f244349c5d6fb0e65cfda4cc3755b/lib/streamlit/elements/lib/form_utils.py#L43-L59

But this makes it so the element after being created has the proper
_form_data, so the above code could just be
```python
return this_dg._active_dg._form_data
```

I have also an alternative suggestion that i think might be much cleaner
but i wanted a second opinion since it would mean deleting 2 tests. But
i find that these tests are highly problematic, let's look:

https://github.com/streamlit/streamlit/blob/4d0013c5b66f244349c5d6fb0e65cfda4cc3755b/lib/tests/streamlit/form_utils_test.py#L62-L75
They are expecting that, if when `_current_form` is called there is a
form in the context stack then it should mean there is a form in the
element being created, it has the same assumption in the `_current_form`
function which iterates the current context until it finds a _form_data
(it does this when the dg is parentless, so st or st.sidebar etc).
However this is wrong and i'll give an example:

```python
no_form = st.empty()
with st.form("form"):
  st.write("hi")
  with no_form:
    st.form_submit_button("OK")
```
In here (run from origin/develop) the form_submit_button is erroneously
thought to be in a form since the `_current_form` function checks the
closest context, sees no form, goes to the next and sees a form, but the
context does not reflect the actual structure, it should iterate the
parents!
I know realistically no one writes an interface like this, but what
happens when you run this piece of code is you get a submit button,
outside of any form, and a form without a submit button inside it,
instead of an exception.

So the cleaner alternative would be:
- Neither _block nor _enqueue propagate _form_data to every element
- Test `test_is_in_form_true_when_dg_stack_has_form` and
`test_is_in_form_false_when_dg_stack_has_no_form` is deleted
- The `_current_form` function actually iterates the parents of the
active_dg until it hits the form dg which has a _form_data
Something like
```python
dg = this_dg._active_dg

while dg is not None:
    if dg._form_data is not None:
        return dg._form_data
    dg = dg._parent
```

Which one works best?

## GitHub Issue Link (if applicable)
Fixes #8761
@Asaurus1 @kmcgrady I'm tagging you two because i saw you were
interested in the issue

## Testing Plan
I have added 2 new unit tests that reflect the desired behavior in the
mentioned issue (these tests fail on the current origin/develop branch)

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.