Fix/sqlite test repeatable (#10889)
## Describe your changes

I noticed that this test is now creating a file on disk and if the test
is run twice in a row it fails since the table/data already exists.
After some research, it seems we need to pass in `uri=True` to ensure
the db is created in-memory.

## GitHub Issue Link (if applicable)

## Testing Plan

No functional changes, test-only. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.