Update placeholder style and in-edit input margin for Selectbox and Multiselect (#10865)
## Describe your changes

## GitHub Issue Link (if applicable)

- Update placeholder color to be the same for Selectbox and Multiselect
- Add `margin-left` to in-edit selectbox to prevent it from jumping to
the left when editing an existing option as shown in the following

## Testing Plan

- The placeholder color change is captured by snapshot tests for which
the snapshots ae updated accordingly
- I guess we could test the style of a selected option vs. one which is
currently edited (the left-jump issue mentioned above tackled by
margin-left) by taking a snapshot of both states and compare them.
However, this is so small of a detail that just for this to have a
snapshot does not sound to be worth it.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.