Fix `showErrorDetails` showing wrong value in `streamlit config show` (#10921)
## Describe your changes

Fixes an issue where the value of `showErrorDetails` is shown as list
and not string in `streamlit config show`:

```
# Default: [ "f", "u", "l", "l",]
# showErrorDetails = [ "f", "u", "l", "l",]
```

This seems to be caused by the value being passed in as Enum and not as
string.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/10913

## Testing Plan

- Add a unit test to ensure this is correct.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.