Migrate issue view counter badge to own provider (#10926)
## Describe your changes

https://github.com/streamlit/streamlit/pull/10922 migrated our view
badge in community voting to https://visitorbadge.io/. However, this
doesn't provide a good way to access the statistics for analysis.
Therefore, I have implemented a very lightweight cloudflare worker using
the cloudlfare key-value store to provide such a feature:

The badge is available via:
```
https://api.views-badge.org/badge/<key>
```

And the stats can be accessed via:
```
https://api.views-badge.org/stats/<key>
```

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.