[RFC] fix chat input height for multi-line placeholder (#10931)
## Describe your changes

See GH issue linked below for context.

In this fix we preserve the scroll height of the input when placeholder
expands it to multiple line, and add new e2e tests to validate 3
scenarios.

However, the caveat is that this fix does not work on Firefox. This is
because when the placeholder text exceeds the predefined height of the
[textarea](https://baseweb.design/components/textarea/), baseweb
increases it scroll height on Chrome and webkit, but not on FIrefox. As
a result, when we calculate `scrollHeight` in `getScrollHeight` it
returns the height set by its styling.


https://github.com/user-attachments/assets/aab34dec-ce39-449c-9a5b-eee01d0af0ee

I'm currently leaning towards shipping this fix, as I think 1) having
this discrepancy has low risk, 2) the existing behavior happens only for
edge cases (when there's a reason for long placeholder or short input
width,) and 3) any other workaround for this baseweb inconsistency feels
hacky and introduces complexity that does not outweigh the gain.

I would appreciate any feedback on how we should go about it. 

## GitHub Issue Link (if applicable)

#10611

## Testing Plan

- E2E Tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.