Use vite proxy for dev server backend connection (#10968)
## Describe your changes

Uses the vite proxy configuration to proxy requests to the Streamlit
backend instead of connecting to a a different port in the frontend.
This is a necessary change to eventually add a devcontainer setup to
this repo: #7242

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.