Only show links to google/ChatGPT if the site is localhost (#10971)
## Describe your changes

We introduced error links with Google/ChatGPT, but we have found that
this does not work well in production scenarios for multiple reasons. We
have decided to only show these links while the url is localhost
(indicating development mode).

## GitHub Issue Link (if applicable)

closes #10924

## Testing Plan

- JS unit tests updated or added.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.