Use github token for downloading develop coverage (#11007)
## Describe your changes

Specifying a run ID in the
[download-artifact](https://github.com/actions/download-artifact)
doesn't seem to be enough, it also requires a github token to access
artifacts from other worklfows. Otherwise, it is just downloading the
artifacts from its own workflow run.

Also added an additional test for hashing since this is the one file
that is different in testing coverage in develop & PRs.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.