Allow auto-reload of config on exceptions (#10857)
## Describe your changes

Fixes an issue where on invalid config in `config.toml` can break the
config watcher and, thereby, breaking the functionality to auto-reload
when the config gets changed.

## Github

- Closes https://github.com/streamlit/streamlit/issues/1256
- Closes https://github.com/streamlit/streamlit/issues/8320 (should
already somewhat work, but with this fix its guaranteed that the
auto-reloading will not break)

## Testing Plan

- Added unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.