Remove st.experimental_audio_input (#11105)
## Describe your changes
Remove `st.experimental_audio_input` widget. 

The removal date was originally `2025-01-01`, but we kept this widget
slightly longer, and now time to remove it.
I checked with the product; the usage of `st.experimental_audio_input`
is around 0, and we can safely remove it now.
Also, there is a hypothesis that LLMs don't like to use `experimental_`
prefixed functionality, so this removal _could make_ st.audio_input more
usable as a side effect.

**This PR is end-to-end vibe coded :) I manually checked that all
required tests for st.audio_input exist, and it works correctly after
those changes.**

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.