Update workflows to support packageManager in package.json (#11119)
## Describe your changes

When I originally did the conversion to yarn, I thought the version of
yarn needed to be specified in CI. I now realize that all that is
necessary is to enable corepack and then the yarn version specified in
packageManager should be downloaded.

## Testing Plan

- Tests should just pass with the correct yarn version

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.