Introduce stableRandomMachineId to metrics by resurrecting old machine_id_v2 code (#11121)
Introduce `stableRandomMachineId` to our metrics by resurrecting old
`machine_id_v2` code. The goal is to keep both stableRandomId and
machineIdVx going forward, to help debug when issues arise.

The logic behind `stableRandomMachineId` is the following:
1. If `~/.streamlit/stable_random_machine_id` exists, use the contents
of that file as the `stableRandomMachineId`.
2. Otherwise, generate `stableRandomMachineId` and store into
`~/.streamlit/stable_random_machine_id`.

And, of course, telemetry is only sent if gatherUsageStats = true. 

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- ✅  Unit Tests (JS and/or Python)
- ~~E2E Tests~~
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.